+++ src/pages/DiscoverPage.tsx
@@ -1,12 +1,14 @@
 import { useInView } from "react-intersection-observer";
 import { useEffect, useRef } from "react";
 import { VIRTUALIZATION } from "@lib/constants";
-import StoryCard, { SkeletonStoryCard } from "@/components/StoryCard";
 import { useDiscoverFeed } from "@hooks/useDiscoverFeed";
 import { useDiscoverFeedVirtualizer } from "@hooks/useDiscoverFeedVirtualizer";
 import DiscoverFeedEmpty from "@components/discover-feed/DiscoverFeedEmpty";
 import DiscoverFeedLoading from "@components/discover-feed/DiscoverFeedLoading";
 import DiscoverFeedError from "@components/discover-feed/DiscoverFeedError";
+import StoryCard, {
+  SkeletonStoryCard,
+} from "@components/discover-feed/StoryCard";
 
 export default function DiscoverPage() {
   const scrollRef = useRef<HTMLDivElement | null>(null);
